FileUpload: fix for empty fileContainer
In this branch, the following error was thrown : 
`Severity: Warning --> Creating default object from empty value /.../vendor/gargron/fileupload/src/FileUpload/FileUpload.php 263`

It was because `$this->fileContainer` was actually empty (not set in any `process()` function).

The solution is to create a File object with the filename, which doesn't corresponds to any actual file. 
Can't do better, since there _is_ an upload error, but not an upload tmp_name value.